requirements.txt
Flask
google-generativeai
gunicorn 
pymongo